+++ src/main/java/springweb/trainingmanager/models/viewmodels/exercise/ExerciseWrite.java
@@ -60,7 +60,7 @@ public class ExerciseWrite extends ExerciseSchema {
         this.repetition = repetition;
     }
 
-    public void setWeights(int weights){
+    public void setWeights(short weights){
         this.weights = weights;
     }
